Build: Add newly found critical -f build options
Both caused unexpected bugs:

-fno-zero-initialized-in-bss
.bss section is not initialized in kernel on hardware. Prevents vars
initialized to zero from being assigned do bss section and not send to bootloader.
Now written to data section.

-fno-delete-null-pointer-checks
It turns out that complier detects null pointer refereces and removes them without warning, when
optimiaztions are enabled. This caused vga driver to crash (compiler cut off half of the function).
Null pointer references are valid in kernel, because we remap page 0 to access devices and it is expected.
Even when it is not valid, we want to detect bugs and trigger segfault, instead of silently generating
unexpected code. (volatile was missing there, but point about broken code is still valid, it didn't even segfault there).
See other discussions:
https://lkml.org/lkml/2018/4/4/601
https://reviews.llvm.org/D47894