VFS: Move to async requests
Giant change,  move to async with some additional nice changes + updates all drivers/fs.
I'm not sure if it is good preformace-wise, but solves problems with syscalls (reissuer) and step to microkernel?
Optional for drivers to operate directly on userspace.
Kernel starts to be more aware and optimal for userspace now.